% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSMSN-package.R
\docType{package}
\name{TSMSN-package}
\alias{TSMSN}
\alias{TSMSN-package}
\title{TSMSN: Truncated Scale Mixtures of Skew-Normal Distributions}
\description{
Return the first four moments, estimation of parameters and sample of the TSMSN distributions (Skew Normal, Skew t, Skew Slash or Skew Contaminated Normal).
}
\author{
\strong{Maintainer}: Eraldo B. dos Anjos Filho \email{ebdaf1@de.ufpe.br} (0000-0002-1272-2794)

Authors:
\itemize{
  \item Aldo M. Garay \email{agaray@de.ufpe.br} (0000-0002-4510-639X)
}

}
\keyword{internal}
